emove genesis's dependence on sequencer
