+++ crates/rooch-genesis/Cargo.toml
@@ -54,7 +54,6 @@ moveos-store = { workspace = true }
 
 rooch-framework = { workspace = true }
 rooch-types = { workspace = true }
-rooch-sequencer = { workspace = true }
 rooch-store = { workspace = true }
 bitcoin-move = { workspace = true }
